+++ lib/cache/index.d.ts
@@ -1,2 +1,1 @@
-export { default } from './cache';
 export * from './cache';