+++ lib/event/event.d.ts
@@ -1,3 +1,1 @@
 export declare function event<T extends Function>(target: Object, propertyKey: string | symbol, descriptor?: TypedPropertyDescriptor<T>): any;
-declare const _default: PropertyDecorator;
-export default _default;