+++ lib/getDecors/index.d.ts
@@ -1,2 +1,1 @@
-export { default } from './getDecors';
 export * from './getDecors';