+++ lib/state/index.d.ts
@@ -1,2 +1,1 @@
-export { default } from './state';
 export * from './state';