+++ lib/state/state.d.ts
@@ -3,4 +3,3 @@ export interface StateValues {
     [key: string]: State;
 }
 export declare function state(target: Object, propertyKey: string, desc?: any): any;
-export default state;