+++ src/getDecors/index.ts
@@ -1,3 +1,2 @@
-export {default} from './getDecors'
 export * from './getDecors'